Refactor builder and vertices.
The builder and vertices have been completely reworked to add
flexibility and group together common functionality.  The Builder is now
responsible for applying a series of network transformations which (1)
can change the connectivity of the network and (2) can change the nature
of the objects or the connections in the network.  Once built the model
is passed to the simulator.

When `run` is called the model is assembled: the IO elements are
introduced (dynamically swapping of function of time nodes where
possible), and the model is passed to the Assembler which performs
one-to-one mappings of intermediate objects to vertices.

Some further neatening is required (objects should probably be
relocated) and unit tests require updating.