Output generator, regions and vertices.
Incomplete.

Initial work required for migration to the new set of SpiNNaker tools.
Tests mostly completed for output generator, completed for regions
(all current types) and incomplete for vertices (themselves incomplete).

The output generator code is reponsible for generating files to write
to memory for regions in subvertices, and generating the app_ptr table
to go with these.